# Xoodoo[12] Permutation Using SSE2 Intrinsics