--- a/example/xoodyak_hash.cpp
+++ b/example/xoodyak_hash.cpp
@@ -13,10 +13,11 @@ main()
   uint8_t* msg = static_cast<uint8_t*>(std::malloc(msg_len));
   uint8_t* out = static_cast<uint8_t*>(std::malloc(xoodyak::DIGEST_LEN));
 
-  random_data(msg, msg_len);
+  xoodyak_utils::random_data(msg, msg_len);
 
   xoodyak::hash(msg, msg_len, out);
 
+  using namespace xoodyak_utils;
   std::cout << "Message         : " << to_hex(msg, msg_len) << std::endl;
   std::cout << "Xoodyak Digest  : " << to_hex(out, xoodyak::DIGEST_LEN)
             << std::endl;