--- a/include/bench/bench_xoodoo.hpp
+++ b/include/bench/bench_xoodoo.hpp
@@ -1,0 +1,26 @@
+#pragma once
+#include "utils.hpp"
+#include "xoodoo.hpp"
+#include <benchmark/benchmark.h>
+
+// Benchmark Xoodyak Authenticated Encryption with Associated Data ( AEAD )
+namespace bench_xoodyak {
+
+// Benchmarks 12 rounds of Xoodoo permutation
+inline void
+xoodoo(benchmark::State& state)
+{
+  uint32_t st[12]{};
+  xoodyak_utils::random_data(st, 12);
+
+  for (auto _ : state) {
+    xoodoo::permute(st);
+
+    benchmark::DoNotOptimize(st);
+    benchmark::ClobberMemory();
+  }
+
+  state.SetBytesProcessed(sizeof(st) * state.iterations());
+}
+
+}