--- a/include/bench/bench_xoodoo.hpp
+++ b/include/bench/bench_xoodoo.hpp
@@ -10,7 +10,7 @@ namespace bench_xoodyak {
 inline void
 xoodoo(benchmark::State& state)
 {
-#if defined __SSE2__
+#if defined __SSE2__ && USE_SSE2 != 0
   alignas(16)
 #endif
     uint32_t st[12]{};