--- a/include/xoodyak.hpp
+++ b/include/xoodyak.hpp
@@ -27,7 +27,11 @@ hash(const uint8_t* const __restrict msg, // N -bytes input message to be hashed
 )
 {
   cyclist::phase_t ph = cyclist::phase_t::Up;
-  uint32_t state[12]{};
+
+#if defined __SSE2__
+  alignas(16)
+#endif
+    uint32_t state[12]{};
 
   cyclist::absorb<cyclist::mode_t::Hash>(state, msg, m_len, &ph);
   cyclist::squeeze<cyclist::mode_t::Hash>(state, out, DIGEST_LEN, &ph);
@@ -50,7 +54,11 @@ encrypt(const uint8_t* const __restrict key,   // 128 -bit secret key
 )
 {
   cyclist::phase_t ph = cyclist::phase_t::Up;
-  uint32_t state[12]{};
+
+#if defined __SSE2__
+  alignas(16)
+#endif
+    uint32_t state[12]{};
 
   cyclist::absorb_key(state, key, nonce, &ph);
   cyclist::absorb<cyclist::mode_t::Keyed>(state, data, dt_len, &ph);
@@ -77,7 +85,11 @@ decrypt(const uint8_t* const __restrict key,   // 128 -bit secret key
 )
 {
   cyclist::phase_t ph = cyclist::phase_t::Up;
-  uint32_t state[12]{};
+
+#if defined __SSE2__
+  alignas(16)
+#endif
+    uint32_t state[12]{};
   uint8_t tag_[16]{};
 
   cyclist::absorb_key(state, key, nonce, &ph);