Adds support for using SSE2 intrinsics when applying Xoodoo[12] permutation. 

For testing SSE2 implementation

```bash
SSE2=1 make
```

For benchmarking SSE2 implementation

```bash
SSE2=1 make benchmark
```

---

Let's look at results

## Without SSE2 using GCC on Intel x86_64

```bash
2023-01-12T07:32:03+00:00
Running ./bench/a.out
Run on (128 X 1027.13 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x64)
  L1 Instruction 32 KiB (x64)
  L2 Unified 1280 KiB (x64)
  L3 Unified 55296 KiB (x2)
Load Average: 0.08, 0.02, 0.01
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
bench_xoodyak::xoodoo                 157 ns          157 ns      4455173 bytes_per_second=291.298M/s
```

## With SSE2 using GCC on Intel x86_64

```bash
2023-01-12T07:32:36+00:00
Running ./bench/a.out
Run on (128 X 1335.66 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x64)
  L1 Instruction 32 KiB (x64)
  L2 Unified 1280 KiB (x64)
  L3 Unified 55296 KiB (x2)
Load Average: 0.11, 0.03, 0.01
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
bench_xoodyak::xoodoo                 123 ns          123 ns      5685346 bytes_per_second=371.848M/s
```

## Without SSE2 using Clang on Intel x86_64

```bash
2023-01-12T07:33:52+00:00
Running ./bench/a.out
Run on (128 X 1349.75 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x64)
  L1 Instruction 32 KiB (x64)
  L2 Unified 1280 KiB (x64)
  L3 Unified 55296 KiB (x2)
Load Average: 0.13, 0.05, 0.01
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
bench_xoodyak::xoodoo                79.1 ns         79.1 ns      8849028 bytes_per_second=578.71M/s
```

## With SSE2 using Clang on Intel x86_64

```bash
2023-01-12T07:34:02+00:00
Running ./bench/a.out
Run on (128 X 1261.26 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x64)
  L1 Instruction 32 KiB (x64)
  L2 Unified 1280 KiB (x64)
  L3 Unified 55296 KiB (x2)
Load Average: 0.18, 0.06, 0.02
-----------------------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------
bench_xoodyak::xoodoo                 116 ns          116 ns      6049088 bytes_per_second=395.593M/s
```